End of Day 1 commit for click handler
Committing work as of end of day 1, click handler is generating the
console log as expected.  And it's taking in the ID value from the DIV
container correctly relating the visualization to the array of the game.
Still need to work out how to pass the updated game status to the backend
API.  That is causing an error.