Click Handler to Console Log
Got click handler to send an API request with the POST game
update for a move.  It only works for X but knows the position and
nothing in there yet can tell if someone won or not. But it updates the game
array on the back end AND the game store I'm going to use for my game
engine. Next step is updating the image when the user clicks.