Finagled picture flip
Now the click handler will flip the picture using a function in
the gameEvent.js file called flipPicture.  It can take in a text variable
of x or o the determine if it should change the blank picture to an x
or an o.  Next step is going to be creating game engine logic to determine
whose turn it is using the gameStore object. That will have to drive both
the value passed to into teh API patch function to update the game status
AND this flipPicture function.