Finished integrating TurnSwitcher
Createad a JS function called turnSwitcher that counts number of x's/o's
in the game cells array to determine whose turn it is.  Integrated that
function with the flipPicture function(...a CALLBACK!) to determine which
picture to flip.  Also added in the JQUERY code that turns off the click
watcher once a user has clicked on an individual cell.