Finalized game status logic
Added in code to the game engine function to check if the gamestore array
has a winning combination or is cats game.  Also added a button to reset
a game once it's started.  The UX of that is very poor and what I will
likely focus on tomorrow.  For now though I think if I post this to
GITPAGES what I have now will meet the requirements.  Made a mistake
doing all this work in MASTER branch, thought I had checked out another
branch.