Fixed glitch with new game button
The new game button was not taking up the entire DIV it was contained in.
That appeard to be causing an odd error where the user could click inside
of the DIV, but not on the button to start a new game.  To resolve it
I just expanded the button and moved it down the page further from the game
board so users won't unintentionally end their games.